moved location of clients
